requirements.txt
grpcio==1.39.0
grpcio-tools==1.39.0
protobuf==3.17.3
six==1.16.0
uuid==1.30
